+++ NEWS.rst
@@ -11,6 +11,7 @@ PyDups 1.1.0 (under development)
   `argcomplete <https://pypi.org/project/argcomplete/>`_.
 * Drop support for Python 3.3 (it is not supported by the 
   `scandir <https://pypi.org/project/scandir/>`_ backport module).
+* Improved compatibility with Python 2.7.
 * Improved `setup.py` (dependency management).
 * Added requirements file.
 * Improved testing: